Move GameSelectorModel, colocate with GameSelectorPanel
