+++ game-core/src/main/java/games/strategy/engine/framework/startup/launcher/LaunchAction.java
@@ -6,7 +6,7 @@ import games.strategy.engine.framework.AutoSaveFileUtils;
 import games.strategy.engine.framework.IGame;
 import games.strategy.engine.framework.LocalPlayers;
 import games.strategy.engine.framework.ServerGame;
-import games.strategy.engine.framework.startup.mc.GameSelectorModel;
+import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import games.strategy.engine.framework.startup.mc.ServerModel;
 import games.strategy.engine.player.Player;
 import java.io.File;
