+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/InGameLobbyWatcherWrapper.java
@@ -1,7 +1,7 @@
 package games.strategy.engine.framework.startup.ui;
 
 import games.strategy.engine.framework.IGame;
-import games.strategy.engine.framework.startup.mc.GameSelectorModel;
+import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import org.triplea.lobby.common.GameDescription.GameStatus;
 
 /**
