+++ game-core/src/main/java/games/strategy/engine/framework/startup/ui/LocalSetupPanel.java
@@ -2,7 +2,7 @@ package games.strategy.engine.framework.startup.ui;
 
 import games.strategy.engine.framework.startup.launcher.ILauncher;
 import games.strategy.engine.framework.startup.launcher.LocalLauncher;
-import games.strategy.engine.framework.startup.mc.GameSelectorModel;
+import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import games.strategy.engine.framework.startup.mc.HeadedLaunchAction;
 import java.util.ArrayList;
 import java.util.List;
