+++ game-core/src/test/java/games/strategy/engine/framework/startup/mc/GameSelectorModelTest.java
@@ -13,6 +13,7 @@ import static org.mockito.Mockito.when;
 
 import games.strategy.engine.data.GameData;
 import games.strategy.engine.data.GameSequence;
+import games.strategy.engine.framework.startup.ui.panels.main.game.selector.GameSelectorModel;
 import games.strategy.engine.framework.ui.GameChooserEntry;
 import games.strategy.triplea.settings.AbstractClientSettingTestCase;
 import java.net.URI;
